refactor: Pull out pins into their own submodule
That way, we don't have pin implementation files and helper files mixed
within the file hierarchy. This makes it easier to navigate the code
base
